Remove redundant Google Map containers & simplify naming
